pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iiif-stitcher"
version = "0.1.0"
description = "Download and stitch IIIF images from NPM digital archive viewer URLs."
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "Pillow>=10.0.0",
]

[project.scripts]
iiif-stitcher = "iiif_stitcher.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
